requirements.txt
Django==1.6.8
GitPython==1.0.1
M2Crypto==0.22.3
MySQL-python==1.2.5
coverage==4.1b2
coveralls==1.1
South==0.8.1
Unipath==1.1
amqp==1.4.6
anyjson==0.3.3
astm==0.5.0
async==0.6.1
billiard==3.3.0.18
celery==3.1.13
django-celery==3.1.10
django-dajax==0.9.2
django-dajaxice==0.6
django-databrowse==1.3
django-debug-toolbar==1.2.1
django-extensions==1.5.5
django-profiler==2.0.0
django-simple-history==1.6.3
django-storages==1.1.8
django-tastypie==0.11.1
factory-boy==2.4.1
fake-factory==0.5.3
geopy==1.1.2
gitdb==0.6.4
gnureadline==6.3.3
ipython==2.2.0
kombu==3.0.21
mimeparse==0.1.3
nose==1.3.4
numpy==1.8.2
pycrypto==2.6.1
pycups==1.9.67
pyodbc==3.0.7
python-dateutil==2.1
python-memcached==1.53
python-mimeparse==0.1.4
python-nmap==0.3.4
pytz==2014.4
requests==2.7.0
selenium==2.42.1
six==1.7.3
smmap==0.9.0
sqlparse==0.1.11
wsgiref==0.1.2

